+++ packages/integrations/src/beszel/beszel-types.ts
@@ -242,3 +242,25 @@ export interface PocketBaseListResponse<T> {
   totalPages: number;
   items: T[];
 }
+
+export interface BeszelSystemRow {
+  id: string;
+  name: string;
+  status: BeszelSystemStatus;
+  cpu: number;
+  memory: number;
+  disk: number;
+  gpu: number;
+  loadAvg: [number, number, number] | null;
+  netBytes: number;
+  temp: number | null;
+  battery: [number, number] | null;
+  services: number;
+  uptime: number;
+  agentVersion: string;
+  hostname: string;
+  cpuModel: string;
+  cores: number;
+  memoryTotal: number;
+  osName: string;
+}
